requirements.txt
dvc
dagshub
scikit-learn
mlflow
dvc-s3 
pandas
ipykernel
numpy
python-dotenv